+++ components/ProfileAvatar/PointsModal/index.tsx
@@ -1,0 +1,1 @@
+export { default as IconAndModal } from "./IconAndModal";
